tests/wpt/tests/html/semantics/menu/tentative/menuitem-interest-invoker.tentative.html

<!doctype html>
<meta charset="utf-8" />
<link rel=author href="mailto:masonf@chromium.org">
<link rel="help" href="https://open-ui.org/components/interest-invokers.explainer">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="../../interestfor/resources/invoker-utils.js"></script>

<div id="interestee"></div>
<menulist id="menu">
  <menuitem id="interestmenuitem" interestfor="interestee">menuitem</menuitem>
</menulist>
<button id="otherbutton">Other Button</button>

<style>
  [interestfor] {
    interest-delay: 0s;
  }
</style>

<script>
  promise_test(async function (t) {
    const interestmenuitem = document.getElementById('interestmenuitem');
    const menu = document.getElementById('menu');
    menu.showPopover(); // So the menuitem is visible for hovering.
    t.add_cleanup(() => {
      menu.hidePopover();
    });

    let event = null;
    interestee.addEventListener("interest", (e) => (event = e));
    await hoverOver(interestmenuitem);
    assert_true(!!event, "InterestEvent is fired");
    assert_true(event instanceof InterestEvent, "event is InterestEvent");
    assert_equals(event.type, "interest", "type");
    assert_false(event.bubbles, "bubbles");
    assert_true(event.composed, "composed");
    assert_true(event.isTrusted, "isTrusted");
    assert_equals(event.target, interestee, "target");
    assert_equals(event.source, interestmenuitem, "source");
  }, "InterestEvent dispatches on menuitem hover");

  promise_test(async function (t) {
    const interestmenuitem = document.getElementById('interestmenuitem');
    const menu = document.getElementById('menu');
    menu.showPopover();

    t.add_cleanup(async () => {
      interestmenuitem.removeAttribute('disabled');
      await hoverOver(otherbutton);
      menu.hidePopover();
    });

    let called = false;
    const listener = () => { called = true; };
    interestee.addEventListener("interest", listener);
    t.add_cleanup(() => interestee.removeEventListener("interest", listener));


    interestmenuitem.setAttribute("disabled", "");
    await hoverOver(interestmenuitem);
    assert_false(called, "event was not called");
  }, "event does not dispatch if invoker menuitem is disabled");
</script>
